feat(#16): 호스트 달력과 게스트 달력 명확하게 구분
